# BUG: Fix weighted rolling variance implementation