--- a/pandas/tests/window/test_win_type.py
+++ b/pandas/tests/window/test_win_type.py
@@ -661,6 +661,21 @@ def test_cmov_window_special_linear_range(win_types_special, step):
     tm.assert_series_equal(xp, rs)
 
 
+@pytest.mark.parametrize("size", [10, 100, 1000])
+@pytest.mark.parametrize("scale", [0, 0.1, 0.01, 0.001, 0.0001])
+def test_weighted_std_through_mean(win_types, step, scale, size):
+    # GH 53273
+    s = Series(np.random.default_rng(0).normal(loc=1, scale=scale, size=size))
+
+    squared_mean = (s**2).rolling(5, win_type=win_types).mean()
+    mean = s.rolling(5, win_type=win_types).mean()
+    expected = (squared_mean - mean**2).pow(0.5)
+
+    result = s.rolling(5, win_type=win_types).std(ddof=0)
+
+    tm.assert_series_equal(expected, result)
+
+
 def test_weighted_var_big_window_no_segfault(win_types, center):
     # GitHub Issue #46772
     pytest.importorskip("scipy")