@mroeschke 2023-09-18
Can you show if there is any performance difference before and after this change?