@ihsansecer 2023-09-18
> Can you show if there is any performance difference before and after this change?

| Change   | Before [ebca6dfb] <main>   | After [a695dc48] <fix-rolling-std-custom-weights>   |   Ratio | Benchmark (Parameter)                           |
|----------|----------------------------|-----------------------------------------------------|---------|-------------------------------------------------|
| +        | 2.18±0.02ms                | 764±4ms                                             |  351.17 | rolling.WeightedVar.time_method(1000, 'boxcar') |
| +        | 2.25±0.02ms                | 768±4ms                                             |  341.83 | rolling.WeightedVar.time_method(1000, 'triang') |
| +        | 2.24±0.03ms                | 73.6±0.3ms                                          |   32.83 | rolling.WeightedVar.time_method(100, 'boxcar')  |
| +        | 2.36±0.3ms                 | 74.6±0.1ms                                          |   31.57 | rolling.WeightedVar.time_method(100, 'triang')  |
| +        | 2.96±0.9ms                 | 5.09±0.06ms                                         |    1.72 | rolling.WeightedVar.time_method(10, 'boxcar')   |
| +        | 3.07±0.9ms                 | 5.07±0.04ms                                         |    1.65 | rolling.WeightedVar.time_method(10, 'triang')   |

---

```python
class WeightedVar:
    params = (
        [10, 100, 1000],
        ["triang", "boxcar"]
    )
    param_names = ["window", "win_type"]

    def setup(self, window, win_type):
        N = 10**5
        arr = np.random.random(N)
        obj = pd.Series(arr)
        self.window = obj.rolling(window, win_type=win_type)

    def time_method(self, window, win_type):
        self.window.var()