@jreback 2023-09-19
the new implementation is O^2 
this is not good at all