@mroeschke 2023-09-27
Is there a way to fix this while keeping the current algorithm? The slowdown is concerning for more common cases where this was already correct