@ihsansecer 2023-10-01
> The scaling and performance at 1000 rows makes this pretty impractical to use for usual datasets I think

@WillAyd It is not the performance at 1000 rows. The number of rows is constant at 100,000 and the varying one is the length of weights. 

> But isn't this algorithm requiring multiple passes instead of simply updating records as the window expands?

It still calculates variance in one pass (over the same value and weight pairs). The problem is nothing here is expanding really. We **have to** recalculate variance for each window because pairs are different in each window. I have an excalidraw board below showing how this works using the same example. Hope this makes sense.

![Untitled-2023-09-30-1837](https://github.com/pandas-dev/pandas/assets/13166114/274ed80e-632b-475b-8dee-88a956445c7b)


By the way, both weighted moving variance and weighted sum have the same complexity O(wn) where w is the number of weights and n is the number of rows.
https://github.com/pandas-dev/pandas/blob/6f0cd8d9ed251aa80f7415c565176fc33487110a/pandas/_libs/window/aggregations.pyx#L1495-L1500